backend/requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
python-dotenv
groq